refactor: rpc namespace refactor - txpool
